# tests/hosts/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "AvoidanceClaim": {
        "additionalProperties": false,
        "properties": {
          "claim_id": {
            "pattern": "^claim_[0-9a-f]{32}$",
            "title": "Claim Id",
            "type": "string"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 64,
            "minItems": 1,
            "title": "Evidence Ids",
            "type": "array"
          },
          "fresh_at": {
            "format": "date-time",
            "title": "Fresh At",
            "type": "string"
          },
          "kind": {
            "const": "avoidance",
            "default": "avoidance",
            "title": "Kind",
            "type": "string"
          },
          "lifecycle": {
            "$ref": "#/components/schemas/ClaimLifecycle",
            "default": "active"
          },
          "value": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "claim_id",
          "confidence",
          "fresh_at",
          "evidence_ids",
          "value"
        ],
        "title": "AvoidanceClaim",
        "type": "object"
      },
...
      "CanonicalProfile": {
        "additionalProperties": false,
        "description": "The sole durable representation of user understanding.",
        "properties": {
          "claims": {
            "default": [],
            "items": {
              "discriminator": {
                "mapping": {
                  "avoidance": "#/components/schemas/AvoidanceClaim",
                  "emerging_interest": "#/components/schemas/EmergingInterestClaim",
                  "insight": "#/components/schemas/InsightClaim",
                  "preference": "#/components/schemas/PreferenceClaim",
                  "stable_interest": "#/components/schemas/StableInterestClaim"
                },
                "propertyName": "kind"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/StableInterestClaim"
                },
                {
                  "$ref": "#/components/schemas/EmergingInterestClaim"
                },
                {
                  "$ref": "#/components/schemas/AvoidanceClaim"
                },
                {
                  "$ref": "#/components/schemas/PreferenceClaim"
                },
                {
                  "$ref": "#/components/schemas/InsightClaim"
                }
              ]
            },
            "maxItems": 500,
            "title": "Claims",
            "type": "array"
          },
          "overrides": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/UserOverride"
            },
            "maxItems": 500,
            "title": "Overrides",
            "type": "array"
          },
          "profile_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Profile Id",
            "type": "string"
          },
          "revision": {
            "minimum": 0.0,
            "title": "Revision",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "profile_id",
          "revision",
          "updated_at"
        ],
        "title": "CanonicalProfile",
        "type": "object"
      },
      "ClaimLifecycle": {
        "enum": [
          "emerging",
          "active",
          "retired"
        ],
        "title": "ClaimLifecycle",
        "type": "string"
      },
...
      "DialogueProfile": {
        "additionalProperties": false,
        "properties": {
          "insights": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Insights",
            "type": "array"
          },
          "preference_summary": {
            "items": {
              "type": "string"
            },
            "maxItems": 30,
            "title": "Preference Summary",
            "type": "array"
          },
          "version": {
            "default": 1,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "preference_summary",
          "insights"
        ],
        "title": "DialogueProfile",
        "type": "object"
      },
...
      "EditProfileResult": {
        "additionalProperties": false,
        "properties": {
          "observation_id": {
            "title": "Observation Id",
            "type": "string"
          },
          "profile": {
            "$ref": "#/components/schemas/CanonicalProfile"
          }
        },
        "required": [
          "profile",
          "observation_id"
        ],
        "title": "EditProfileResult",
        "type": "object"
      },
      "EmergingInterestClaim": {
        "additionalProperties": false,
        "properties": {
          "claim_id": {
            "pattern": "^claim_[0-9a-f]{32}$",
            "title": "Claim Id",
            "type": "string"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 64,
            "minItems": 1,
            "title": "Evidence Ids",
            "type": "array"
          },
          "fresh_at": {
            "format": "date-time",
            "title": "Fresh At",
            "type": "string"
          },
          "kind": {
            "const": "emerging_interest",
            "default": "emerging_interest",
            "title": "Kind",
            "type": "string"
          },
          "lifecycle": {
            "$ref": "#/components/schemas/ClaimLifecycle",
            "default": "emerging"
          },
          "value": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "claim_id",
          "confidence",
          "fresh_at",
          "evidence_ids",
          "value"
        ],
        "title": "EmergingInterestClaim",
        "type": "object"
      },
...
      "ErrorCode": {
        "enum": [
          "validation",
          "unauthorized",
          "forbidden",
          "unavailable_capability",
          "conflict",
          "rate_limit",
          "temporary_failure",
          "not_found",
          "method_not_allowed"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "maxLength": 500,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "InsightClaim": {
        "additionalProperties": false,
        "properties": {
          "claim_id": {
            "pattern": "^claim_[0-9a-f]{32}$",
            "title": "Claim Id",
            "type": "string"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 64,
            "minItems": 1,
            "title": "Evidence Ids",
            "type": "array"
          },
          "fresh_at": {
            "format": "date-time",
            "title": "Fresh At",
            "type": "string"
          },
          "kind": {
            "const": "insight",
            "default": "insight",
            "title": "Kind",
            "type": "string"
          },
          "lifecycle": {
            "$ref": "#/components/schemas/ClaimLifecycle",
            "default": "active"
          },
          "value": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "claim_id",
          "confidence",
          "fresh_at",
          "evidence_ids",
          "value"
        ],
        "title": "InsightClaim",
        "type": "object"
      },
...
      "OverrideOperation": {
        "enum": [
          "set",
          "remove"
        ],
        "title": "OverrideOperation",
        "type": "string"
      },
...
      "PreferenceClaim": {
        "additionalProperties": false,
        "properties": {
          "claim_id": {
            "pattern": "^claim_[0-9a-f]{32}$",
            "title": "Claim Id",
            "type": "string"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "dimension": {
            "$ref": "#/components/schemas/PreferenceDimension"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 64,
            "minItems": 1,
            "title": "Evidence Ids",
            "type": "array"
          },
          "fresh_at": {
            "format": "date-time",
            "title": "Fresh At",
            "type": "string"
          },
          "kind": {
            "const": "preference",
            "default": "preference",
            "title": "Kind",
            "type": "string"
          },
          "lifecycle": {
            "$ref": "#/components/schemas/ClaimLifecycle",
            "default": "active"
          },
          "value": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "claim_id",
          "confidence",
          "fresh_at",
          "evidence_ids",
          "dimension",
          "value"
        ],
        "title": "PreferenceClaim",
        "type": "object"
      },
      "PreferenceDimension": {
        "enum": [
          "content",
          "style",
          "creator",
          "language",
          "provider"
        ],
        "title": "PreferenceDimension",
        "type": "string"
      },
...
      "ProfileEditRequest": {
        "additionalProperties": false,
        "properties": {
          "account_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Account Id",
            "type": "string"
          },
          "claim_id": {
            "pattern": "^claim_[0-9a-f]{32}$",
            "title": "Claim Id",
            "type": "string"
          },
          "idempotency_key": {
            "maxLength": 200,
            "minLength": 8,
            "title": "Idempotency Key",
            "type": "string"
          },
          "operation": {
            "enum": [
              "set",
              "remove"
            ],
            "title": "Operation",
            "type": "string"
          },
          "profile_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Profile Id",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "idempotency_key",
          "profile_id",
          "account_id",
          "claim_id",
          "operation"
        ],
        "title": "ProfileEditRequest",
        "type": "object"
      },
      "ProfileResponse": {
        "additionalProperties": false,
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/DialogueProfile"
          }
        },
        "required": [
          "profile"
        ],
        "title": "ProfileResponse",
        "type": "object"
      },
...
      "StableInterestClaim": {
        "additionalProperties": false,
        "properties": {
          "claim_id": {
            "pattern": "^claim_[0-9a-f]{32}$",
            "title": "Claim Id",
            "type": "string"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 64,
            "minItems": 1,
            "title": "Evidence Ids",
            "type": "array"
          },
          "fresh_at": {
            "format": "date-time",
            "title": "Fresh At",
            "type": "string"
          },
          "kind": {
            "const": "stable_interest",
            "default": "stable_interest",
            "title": "Kind",
            "type": "string"
          },
          "lifecycle": {
            "$ref": "#/components/schemas/ClaimLifecycle",
            "default": "active"
          },
          "value": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "claim_id",
          "confidence",
          "fresh_at",
          "evidence_ids",
          "value"
        ],
        "title": "StableInterestClaim",
        "type": "object"
      },
...
      "UserOverride": {
        "additionalProperties": false,
        "properties": {
          "claim_id": {
            "pattern": "^claim_[0-9a-f]{32}$",
            "title": "Claim Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "operation": {
            "$ref": "#/components/schemas/OverrideOperation"
          },
          "override_id": {
            "pattern": "^override_[0-9a-f]{32}$",
            "title": "Override Id",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "override_id",
          "claim_id",
          "operation",
          "created_at"
        ],
        "title": "UserOverride",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OpenBiliClaw API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/profiles/edit": {
      "post": {
        "operationId": "edit_profile_v1_profiles_edit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileEditRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditProfileResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not_found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "method_not_allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "rate_limit"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "temporary_failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unavailable capability or temporary timeout"
          }
        },
        "summary": "Edit Profile",
        "tags": [
          "understanding"
        ]
      }
    },
    "/v1/profiles/{profile_id}": {
      "get": {
        "operationId": "profile_v1_profiles__profile_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not_found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "method_not_allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "rate_limit"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "temporary_failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unavailable capability or temporary timeout"
          }
        },
        "summary": "Profile",
        "tags": [
          "understanding"
        ]
      }
    },
...
  }
}